requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via google-genai
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   google-genai
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
azure-core==1.35.0
    # via azure-storage-blob
azure-storage-blob==12.26.0
    # via -r requirements.in
backoff==2.2.1
    # via posthog
bcrypt==4.3.0
    # via chromadb
build==1.3.0
    # via chromadb
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   kubernetes
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.3
    # via requests
chroma-haystack==3.3.0
    # via -r requirements.in
chromadb==1.0.13
    # via
    #   -r requirements.in
    #   chroma-haystack
click==8.2.1
    # via
    #   typer
    #   uvicorn
coloredlogs==15.0.1
    # via onnxruntime
cryptography==45.0.6
    # via azure-storage-blob
distro==1.9.0
    # via
    #   openai
    #   posthog
docstring-parser==0.17.0
    # via
    #   haystack-ai
    #   haystack-experimental
durationpy==0.10
    # via kubernetes
et-xmlfile==2.0.0
    # via openpyxl
fastapi==0.116.1
    # via -r requirements.in
filelock==3.19.1
    # via huggingface-hub
filetype==1.2.0
    # via
    #   haystack-ai
    #   haystack-experimental
flatbuffers==25.2.10
    # via onnxruntime
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.7.0
    # via huggingface-hub
google-auth==2.40.3
    # via
    #   google-genai
    #   kubernetes
google-genai[aiohttp]==1.39.1
    # via google-genai-haystack
google-genai-haystack==2.1.2
    # via -r requirements.in
googleapis-common-protos==1.70.0
    # via opentelemetry-exporter-otlp-proto-grpc
grpcio==1.74.0
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
haystack-ai==2.17.1
    # via
    #   -r requirements.in
    #   chroma-haystack
    #   google-genai-haystack
    #   haystack-experimental
haystack-experimental==0.12.0
    # via haystack-ai
hf-xet==1.1.8
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   chromadb
    #   google-genai
    #   openai
huggingface-hub==0.34.4
    # via tokenizers
humanfriendly==10.0
    # via coloredlogs
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via opentelemetry-api
importlib-resources==6.5.2
    # via chromadb
isodate==0.7.2
    # via azure-storage-blob
jinja2==3.1.6
    # via haystack-ai
jiter==0.10.0
    # via openai
jsonref==1.1.0
    # via google-genai-haystack
jsonschema==4.25.1
    # via
    #   chromadb
    #   haystack-ai
jsonschema-specifications==2025.4.1
    # via jsonschema
kubernetes==33.1.0
    # via chromadb
lazy-imports==1.0.1
    # via haystack-ai
lxml==6.0.1
    # via
    #   python-docx
    #   python-pptx
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mmh3==5.2.0
    # via chromadb
more-itertools==10.7.0
    # via haystack-ai
mpmath==1.3.0
    # via sympy
multidict==6.6.4
    # via
    #   aiohttp
    #   yarl
networkx==3.5
    # via haystack-ai
numpy==2.3.2
    # via
    #   chromadb
    #   haystack-ai
    #   onnxruntime
    #   pandas
oauthlib==3.3.1
    # via
    #   kubernetes
    #   requests-oauthlib
onnxruntime==1.22.1
    # via chromadb
openai==1.100.2
    # via haystack-ai
openpyxl==3.1.5
    # via -r requirements.in
opentelemetry-api==1.36.0
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp-proto-common==1.36.0
    # via opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-grpc==1.36.0
    # via chromadb
opentelemetry-proto==1.36.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-sdk==1.36.0
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-semantic-conventions==0.57b0
    # via opentelemetry-sdk
orjson==3.11.2
    # via chromadb
overrides==7.7.0
    # via chromadb
packaging==25.0
    # via
    #   build
    #   huggingface-hub
    #   onnxruntime
pandas==2.3.2
    # via -r requirements.in
pillow==11.3.0
    # via python-pptx
posthog==6.6.0
    # via
    #   chromadb
    #   haystack-ai
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
protobuf==6.32.0
    # via
    #   googleapis-common-protos
    #   onnxruntime
    #   opentelemetry-proto
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pybase64==1.4.2
    # via chromadb
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   chromadb
    #   fastapi
    #   google-genai
    #   haystack-ai
    #   openai
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via rich
pypika==0.48.9
    # via chromadb
pyproject-hooks==1.2.0
    # via build
python-dateutil==2.9.0.post0
    # via
    #   haystack-ai
    #   kubernetes
    #   pandas
    #   posthog
python-docx==1.2.0
    # via -r requirements.in
python-dotenv==1.1.1
    # via uvicorn
python-multipart==0.0.20
    # via -r requirements.in
python-pptx==1.0.2
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   chromadb
    #   haystack-ai
    #   huggingface-hub
    #   kubernetes
    #   uvicorn
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   -r requirements.in
    #   azure-core
    #   google-genai
    #   haystack-ai
    #   huggingface-hub
    #   kubernetes
    #   posthog
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via kubernetes
rich==14.1.0
    # via
    #   chromadb
    #   typer
rpds-py==0.27.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
schedule==1.2.2
    # via -r requirements.in
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   azure-core
    #   kubernetes
    #   posthog
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
starlette==0.47.2
    # via fastapi
sympy==1.14.0
    # via onnxruntime
tenacity==9.1.2
    # via
    #   chromadb
    #   google-genai
    #   haystack-ai
tokenizers==0.21.4
    # via chromadb
tqdm==4.67.1
    # via
    #   chromadb
    #   haystack-ai
    #   huggingface-hub
    #   openai
typer==0.16.1
    # via chromadb
typing-extensions==4.14.1
    # via
    #   aiosignal
    #   anyio
    #   azure-core
    #   azure-storage-blob
    #   chromadb
    #   fastapi
    #   google-genai
    #   haystack-ai
    #   huggingface-hub
    #   openai
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   posthog
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   python-pptx
    #   referencing
    #   starlette
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   kubernetes
    #   requests
uvicorn[standard]==0.35.0
    # via
    #   -r requirements.in
    #   chromadb
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websocket-client==1.8.0
    # via kubernetes
websockets==15.0.1
    # via
    #   google-genai
    #   uvicorn
xlsxwriter==3.2.5
    # via python-pptx
yarl==1.20.1
    # via aiohttp
zipp==3.23.0
    # via importlib-metadata
